+++ tests/unit/test_destroy_api.py
@@ -62,7 +62,7 @@ POST_DESTROY_ACTIONS: dict[str, Callable[[WebView], object]] = {
     "rpc": lambda w: w.rpc(_ping),
     "unexpose": lambda w: w.unexpose("ping"),
     "watch_app": lambda w: w.watch_app(),
-    "set_on_navigation": lambda w: w.set_on_navigation(lambda _u: True),
+    "set_on_navigation": lambda w: w.set_on_navigation(lambda _event: True),
     "set_navigation_policy": lambda w: w.set_navigation_policy(lambda _e: True),
     "set_on_page_load": lambda w: w.set_on_page_load(lambda *_a: None),
     "set_on_title_changed": lambda w: w.set_on_title_changed(lambda _t: None),